requirements.txt
Django==4.1.5
djangorestframework==3.14.0
django-kronos==1.0
django-environ==0.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
sentry-sdk==1.13.0
django-translated-fields==0.12.0
Pillow==9.4.0
requests==2.28.2
django-admin-sortable2==2.1.4
django-cleanup==6.0.0
django-sslserver==0.22.0
django-dump-load-utf8==0.0.4
black==22.12.0
flake8==6.0.0
pre-commit==2.21.0
PyJWT==2.6.0
requests-oauthlib==1.3.1
django-colorfield==0.8.0
openai==0.26.0
django-admin-interface==0.24.2
